requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==23.2.0
backports.weakref==1.0.post1
blinker==1.7.0
brotlipy==0.7.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
constantly==15.1.0
cssselect==1.1.0
dateparser==1.2.0
debugpy==1.8.1
decorator==5.1.1
docstring-to-markdown==0.13
et-xmlfile==1.1.0
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.25.0
frozenlist==1.4.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.4
multipledispatch==1.0.0
munkres==1.1.4
mypy-extensions==0.4.3
namex==0.0.7
nest-asyncio==1.6.0
numpy==1.26.2
openpyxl==3.0.10
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pep8==1.7.1
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.3.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==15.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.19.0
pycurl==7.45.2
pydeck==0.8.1b0
PyDispatcher==2.0.5
Pygments==2.17.2
pyls-spyder==0.4.0
PyQt5_sip==12.11.0
pyspark==3.5.1
python-binance==1.0.19
python-dateutil==2.8.2
python-lsp-jsonrpc==1.1.0
python-lsp-server==1.9.0
pytz==2023.3.post1
pyzmq==26.0.0
queuelib==1.5.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.2.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.24
stack-data==0.6.3
streamlit==1.32.2
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.9.0
tzlocal==5.2
ucimlrepo==0.0.2
ujson==5.9.0
urllib3==2.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
